feat: Add inventory kitchen equipment query
